fixed all files

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -1,10 +1,13 @@
+#include <stdlib.h>
 #include "lists.h"
+
 /**
- * list_len - Finds the number of elements in a linked list
- * @h: pointer to the list
- * Return: the number of elements
+ * list_len - returns the number of elements in a linked list
+ * @h: pointer to the list_t list
+ *
+ * Return: number of elements in h
  */
-int list_len(const list_t *h)
+size_t list_len(const list_t *h)
 {
 	size_t count = 0;
